lab-building-agenticai-with-agentcore/requirements.txt
strands-agents
strands-agents-tools
boto3>=1.40.8
botocore>=1.40.8
bedrock-agentcore>=0.1.2
bedrock-agentcore-starter-toolkit>=0.1.5
aws-opentelemetry-distro
ddgs
aws-opentelemetry-distro~=0.10.1
pyyaml
typing_extensions>=4.0.0